requirements.txt
albucore==0.0.21
albumentations==1.4.23
annotated-types==0.7.0
eval_type_backport==0.2.0
filelock==3.16.1
fsspec==2024.10.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python-headless==4.10.0.84
pillow==11.0.0
pydantic==2.10.3
pydantic_core==2.27.1
PyYAML==6.0.2
scipy==1.13.1
simsimd==6.2.1
stringzilla==3.11.1
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
